// ui/src/gen/entities/organization_pb.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=ts"
// @generated from file entities/organization.proto (package process, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file entities/organization.proto.
 */
export const file_entities_organization: GenFile = /*@__PURE__*/
  fileDesc("ChtlbnRpdGllcy9vcmdhbml6YXRpb24ucHJvdG8SB3Byb2Nlc3MiPQoMT3JnYW5pemF0aW9uEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAlCkwEKC2NvbS5wcm9jZXNzQhFPcmdhbml6YXRpb25Qcm90b1ABWjVnaXRodWIuY29tL2dzb3VsdGFuL2dvYnBtL2FwaS9wcm90by9lbnRpdGllcztlbnRpdGllc6ICA1BYWKoCB1Byb2Nlc3PKAgdQcm9jZXNz4gITUHJvY2Vzc1xHUEJNZXRhZGF0YeoCB1Byb2Nlc3NiBnByb3RvMw");

/**
 * @generated from message process.Organization
 */
export type Organization = Message<"process.Organization"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: string description = 3;
   */
  description: string;
};

/**
 * Describes the message process.Organization.
 * Use `create(OrganizationSchema)` to create a new message.
 */
export const OrganizationSchema: GenMessage<Organization> = /*@__PURE__*/
  messageDesc(file_entities_organization, 0);